The App design will be based on the Facade design pattern. The main facade is SuitAlterations project and also it's entry-point of this app.
1. Add SuitAlterations.ServiceBusTopic as encapsulated app layer which will be responsible for:
- handling of the Azure Topic messages;
- sending notification to the Core App layer (see pt. 2) for the further its processing with saving into db
This app layer shouldn't have any references on other projects and we can communicate with other app layers by sending notifications using MediatR pattern and self-titled nuget package.

Register our azure topic message handler in Startup class, after that our handler will start to listen to an incoming messages in Azure Topic.
In addition, we need to add an app settings file which will contain the Azure Service Bus settings (connection string, topic, and subscription names and etc.). In the Development env. we need to use user secrets, this is allowed to us do not commit and store the real settings to VCS.

Moreover, we need to wrap all callings to the MediatR instance into our notifications service which will send PaidSuitAlteration notification when the alteration has been paid.

1.1. The next part of working with the Azure topic and Subscriptions is covering of the logic by Integration tests. To achieve this aim we need to add new project with integration tests for SuitAlterations.ServiceBusTopic.
First of all, we need to add setup method which will create test PaidAlteration message for us, and in this way, we can fully test our logic for the handling Azure Topic Subscription messages. Just for now it will be the single test and it will be ignored by default.
Since we no need to test logic with the further processing notification here because it should be covered by tests within the App Core layer, then we need to mock it here. Also, we need to make sure here that PaidAlteration message was sent after handling by the SuitAlterationTopicSubscription and that's it.

2. As was mentioned above, we need to add one more app layer which will represent the Core server layer of the app. We need to add SuitAlterationsService which will process notification message with the suit alteration id and update its status in db. Just for now it contains the stub for that. It will be implemented in the next commit