Prerequisites for the below changes:  - create Azure SQL Db  - add an according appsetting which will contain connectionstring to Azure sql db;  - add nuget packages to work with ef core sql and ef core tools accordingly
The next part of the SuitAlterations app is the Core layer where we need to provide service which responsible for managing the customers and their suit alterations (create/read/update), including an updating the suit alterations status to Paid after we got PaidSuitAlteration notification from the Azure Topic Subscription.

But first of all, Create ApplicationDbContext class, Customer and SuitAlteration entities and also reps for them. Also wrap the work with dbcontext into GenericRepository and create custom for suitalterations to get alterations which are filtered by customer id.

Add enum with possible Suit alteration statuses - create, paid and done

Use the Code-First mechanism with Db migrations

Cover logic in the SuitAlterationsService by unit tests, but in fact, we won't mock here our DbContext and its reps, we will use the In-memory database which easily can be configured for the test needs.