The next layer in our architecture is Infrastructure layer in accordance with the Clean Architecture.
--
Configuration within this layer includes the ApplicationDbContext to work with DB.
DB migrations are consolidated in this layer as well.
Since we should get messages from the Azure Service Bus Topic, we need to provide some configuration model (AzureServiceBusTopicSubscriptionConfiguration.cs) which we can use to get access to our Service Bus Topic and its subscription.

--
Because we should stick to the DDD approach, then all entities will be setting up by using Fluent API to map Entities and their properties and Entity Relationships to database.
We can't use Data Annotations because it violates DDD principles.
Due to this fact, we use entity type configurations which are implements IEntityTypeConfiguration<TEntity>.
The relation between Customer and its orders (SuitAlterations) is configured by using Foreign key shadow property (SuitAlterations).

--
We have entity repositories that are responsible to work with DbContext directly to get something data from DB.
Btw, these entity reps have own base GenericRepository.cs with base methods to exchange data.
--

To handle domain events we introduced DomainEventsDispatcher.cs which is dispatches such events. The current dispatcher uses MediatR under the hood to publish these events for its subscribers

--

UoW
The Unit of Work sits in the Infrastructure layer and will only focus on transaction management, plus it's responsible for the triggering dispatching of events.
Every change is accompanied by starting transaction scope and disposing after completing it.
Therefore the Unit of Work starts at the beginning of the use-case and finishes at the end of the use-case, and all this happens inside the Application Services (Application layer).

--

After research the implementation way for receiving "OrderPaid" messages in Azure topic - I decided that we need to add some correlation filter (OrderPaidMessageFilter.cs) and thereby we will say, that our OrderPaidMessageReceiver will receive only messages of this type. Otherwise, it can become the common receiver of any messages but is wrong in our case.

--

The last one is ApplicationStartup.cs, which is responsible for registering app services and some configuration. The another one responsibility is registering service for receiving messages from the Azure Service Bus Topic in the application pipeline.