The outside layer in our app is API which is bordering on the same level with the Infrastructure layer and which is sharing registrations of some services within this layer as well.
The API layer can interact directly only with the Infrastructure layer or with the Application layer through Infrastructure

Added some UI adjustments in accordance with changes and relations between application layers

To keep layers persistence we had a need to encapsulate domain Identity objects within the Application layer, thereby, we had started using primitive type instead of a Value Object type for the entity identities

extract a registration of the Automapper profile from API to the Infrastructure layer. The same to do for MediatR requests.

minor fix with initialization of customer's SuitAlterations collection